Reorganize App components into App, AppDND and AppUI
